Add functionality to pause, add break timer after timer gets complete
The break timer executes once the main timer gets over and the user is asked for a popup whether he wants to take a break or not(execute the break timer or not), if called yes then the break timer gets executed and the state is set to the break timer state. There on he can manually start the timer. Once the break timer is finished, the user won't be asked for any input and the state would be automatically changed to the initial timer state. The same process can happen again now.
The user can single tap on the timer (the text) to pause/resume the timer and double tap the timer (in the reset state) to start the timer.